Add decouple model from controller by service and IoC  - The db can use long for id and controller can use int for requests.

+++ conf/evolutions/default/1.sql
@@ -3,7 +3,7 @@
 # --- !Ups
 ---
 create table restaurant (
-  id                        int not null,
+  id                        bigint not null,
   name                      varchar(255) not null,
   address1                  varchar(255) not null,
   address2                  varchar(255) not null,
